fix(MachinePool): always update vmssState
Updates ScaleSets service to always update the scope's vmss state with
the latest data.
Previously, a long running operation would cause the vmssState to not be
updated and delayed creation of AzureMachinePoolMachines until much
later when that long running operation completed.
This change avoids this and updates the vmssState to what is retrieved
from the API at all times.
The call does not make more API requests because the VMSS GET request
was done anyway but its result ignored.